+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 25.11
+
+* The following resources have been added
+
+  * GET `/users/me/agents/queues`
+
 ## 25.10
 
 * The following resources now expose the `queues` field containing per queue status for agents